pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wallpaper-manager"
version = "0.1.1"
description = "Wallpaper Engine 本地壁纸库管理工具"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "PySide6>=6.6.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
video = [
    "PySide6-Multimedia>=6.6.0",
]
dev = [
    "pyinstaller>=6.0",
    "ruff",
]

[project.scripts]
wallpaper-manager = "app:main"

[tool.setuptools.packages.find]
include = ["core*", "ui*"]

[tool.ruff]
line-length = 100
target-version = "py310"
